Update wordpress-quiz.md
I have suggest answers to some question that did not have answers.

ALSO
#### Q25. You might see this code in a WordPress plugin. What does it do?
I think the answer is wrong.

https://wordpress.stackexchange.com/questions/319407/where-i-should-not-use-if-definedabspath-exit

```The point of that code is to prevent any PHP inside the file from being executed if the file is accessed directly outside a WordPress context. ABSPATH is defined by WordPress, so if it's missing when the file is accessed you can tell that it's not running in a WordPress context. So the only place you wouldn't use it is in any file that you did need to access directly. There probably shouldn't be any such files in a WordPress theme or plugin. The most common example might be a file for handling AJAX requests, but in WordPress you should be using admin-ajax.php or the REST API for that sort of thing.```

I believe it should be this   ```[ ] This is a security measure. ABSPATH is the absolute path to the WordPress directory. If the file is called directly, ABSPATH will not be defined and therefore the plugin should not run.```

anyone care to way in ?

- - -
#### Q21. The REST API provides public data, which is accessible to any client anonymously, as well as private data available only after authentication. How could you ensure that no one can anonymously access site data via the REST API?

https://themeisle.com/blog/wordpress-disable-xmlrpc/
- - -
#### Q61. Which of these CSS classs naming convention is correct according to WordPress CSS Coding Standards?
Incorrect answer
updated answer to
- [x] `.selector-name`
https://developer.wordpress.org/coding-standards/wordpress-coding-standards/css/
- - -